fix: Move Post component
