+++ src/components/PostPage.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { Post } from '@/components/Post';
+import { Post } from '@/components/Post/Post';
 //import { IPost } from 'interfaces';
 import { NextSeo } from 'next-seo';
